适配部分在热更层注册的自定义json扩展

+++ Assets/Scripts/Hotfix/TestJson1_Data.cs
@@ -8,7 +8,7 @@ public class TestJson1_Root
 {
     
     public bool b;
-    public float num;
+    public double num;
     public string str;
     public List<int> intList;
     public Dictionary<string, int> intDict;
@@ -20,7 +20,7 @@ public class TestJson1_Root
 public class TestJson1_Item
 {
     public bool b;
-    public float num;
+    public double num;
     public string str;
 
     public override string ToString()
